more int overflow risks (#1100)

+++ poi/src/main/java/org/apache/poi/hssf/record/FeatHdrRecord.java
@@ -103,7 +103,7 @@ public final class FeatHdrRecord extends StandardRecord {
 
         out.writeShort(isf_sharedFeatureType);
         out.writeByte(reserved);
-        out.writeInt((int)cbHdrData);
+        out.writeInt(Math.toIntExact(cbHdrData));
         out.write(rgbHdrData);
     }
